backend/requirements.txt
flask
flask-cors
pymongo[srv]
dnspython
python-dotenv
requests

deep-translator
langdetect
gTTS
pydub
SpeechRecognition
pillow
numpy
scikit-learn
joblib
faiss-cpu
gunicorn
google-generativeai